+++ cobasi/src/main/java/br/edu/insper/desagil/pi/cobasi/Historico.java
@@ -2,45 +2,16 @@ package br.edu.insper.desagil.pi.cobasi;
 
 import java.time.LocalDate;
 
-public class Historico {
-    private String nomeTutor;
-    private String enderecoTutor;
-    private String telefoneTutor;
-    private String nome;
-    private int tipo;
-    private LocalDate dataNascimento;
-    private double peso;
-    private String extra;
 
-    public Historico(String nomeTutor, String nome, int tipo, LocalDate dataNascimento, double peso, String extra) {
-        this.nomeTutor = nomeTutor;
-        this.enderecoTutor = null;
-        this.telefoneTutor = null;
-        this.nome = nome;
-        this.tipo = tipo;
-        this.dataNascimento = dataNascimento;
-        this.peso = peso;
-        this.extra = extra;
+public class Historico extends Animal {
+    private Tutor tutor;
+    private Animal animal;
+
+    public Historico(Tutor tutor, String nome, int tipo, LocalDate dataNascimento, double peso, String extra) {
+        this.tutor = tutor;
+        super(nome, tipo, dataNascimento, peso, extra);
     }
 
-    public void setEnderecoTutor(String enderecoTutor) {
-        this.enderecoTutor = enderecoTutor;
-    }
-
-    public void setTelefoneTutor(String telefoneTutor) {
-        this.telefoneTutor = telefoneTutor;
-    }
-
-    public String resumoTutor() {
-        String summary = "Tutor: " + nomeTutor + "\n";
-        if (enderecoTutor != null) {
-            summary += "Endereço: " + enderecoTutor + "\n";
-        }
-        if (telefoneTutor != null) {
-            summary += "Telefone: " + telefoneTutor + "\n";
-        }
-        return summary;
-    }
 
     public String resumo() {
         String summary = "Animal:" + nome + "\n";
